requirements.txt
streamlit==1.28.1
spotipy==2.23.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.1
streamlit-option-menu==0.3.6
streamlit-authenticator==0.2.3
streamlit-aggrid==0.3.4
streamlit-player==0.1.5 